src/gui/requirements.txt
Pillow==9.0.1
PySide6==6.2.3
shiboken6==6.2.3
